requirements.txt
Flask==1.0.2
Flask_Script==2.0.6
Flask_Cache==0.13.1
Flask_Migrate==2.2.1
Flask_SQLAlchemy==2.3.2
lxml==4.2.3
qrcode==6.0
redis==2.10.6
requests==2.19.1
SQLAlchemy==1.2.10
Flask_Cors==3.0.6
Werkzeug==0.14.1
celery==4.2.1
pycrypto==2.6.1
Pillow==5.4.1
gevent==1.3.5
greenlet==0.4.14
gunicorn==19.9.0
